The main intention to propose this idea "Food Bridge" is every day, restaurants, grocery stores, and event organizers throw away perfectly good food — while nearby families struggle with hunger. (#118)
The problem isn’t the lack of food, but the lack of connection. There’s no simple, centralized system to link food donors with charities and shelters that can use this food quickly and safely.

FoodBridge aims to solve this by providing a ServiceNow-powered platform that instantly connects food donors with organizations in need, helping reduce waste and fight hunger at the same time.so I proposed this Idea